fix infinityScroll styles & change directory for RegisteredUsersCounter
