+++ src/features/posts/ui/postList/postList.module.scss
@@ -18,3 +18,7 @@
     height: 100%;
   }
 }
+
+.infiniteScrollWrapper {
+  overflow: hidden !important;
+}
