+++ src/features/public/ui/index.ts
@@ -1,0 +1,3 @@
+export * from './publicPost'
+export * from './publicPosts'
+export * from './registeredUsersCounter'
