+++ src/features/public/ui/publicPost/index.ts
@@ -1,0 +1,1 @@
+export * from './publicPost'
