+++ src/features/public/ui/publicPost/publicPost.module.scss
@@ -1,0 +1,4 @@
+.user{
+  display: flex;
+  gap: 10px;
+}
