+++ src/features/public/ui/publicPosts/index.ts
@@ -1,0 +1,1 @@
+export * from './publicPosts'
