+++ src/features/public/ui/publicPosts/publicPosts.module.scss
@@ -9,4 +9,6 @@
 
 .posts{
   display: flex;
+  gap: 11px;
+  max-height: 391px;
 }
