+++ src/features/public/ui/registeredUsersCounter/index.ts
