+++ src/features/public/ui/registeredUsersCounter/registeredUsersCounter.module.scss
@@ -4,6 +4,7 @@
   justify-content: space-between;
 
   width: 100%;
+  margin: 0 0 20px;
   padding: 24px;
 }
