+++ src/features/public/ui/registeredUsersCounter/registeredUsersCounter.tsx
