+++ src/pages/index.tsx
@@ -1,13 +1,7 @@
-import { GetPublicPostsResponse } from '@/features'
-import { API_URLS, ROUTES } from '@/shared/config'
-import { ProfileAvatar, RegisteredUsersCounter } from '@/shared/ui'
+import { GetPublicPostsResponse, PublicPosts, RegisteredUsersCounter } from '@/features'
+import { API_URLS } from '@/shared/config'
 import { GetStaticProps } from 'next'
 import Head from 'next/head'
-import Image from 'next/image'
-import Link from 'next/link'
-import { useRouter } from 'next/router'
-
-import styles from '@/features/public/publicPage.module.scss'
 
 export const getStaticProps: GetStaticProps = async () => {
   const res = await fetch(
@@ -19,24 +13,6 @@ export const getStaticProps: GetStaticProps = async () => {
 }
 
 const Home = ({ items, totalUsers }: Pick<GetPublicPostsResponse, 'items' | 'totalUsers'>) => {
-  const router = useRouter()
-
-  const handlePostClick = (ownerId: number, postId: number) => {
-    router.push(
-      {
-        pathname: `${ROUTES.PROFILE}/${ownerId}`,
-        query: { postId: postId.toString() },
-      },
-      undefined,
-      { shallow: true }
-    )
-  }
-
-  const classNames = {
-    container: styles.container,
-    posts: styles.posts,
-  } as const
-
   return (
     <>
       <Head>
@@ -48,31 +24,9 @@ const Home = ({ items, totalUsers }: Pick<GetPublicPostsResponse, 'items' | 'tot
         <meta content={`public, photo, fiesta, github`} name={'keywords'} />
         <meta content={'index, follow'} name={'robots'} />
       </Head>
-      <div className={classNames.container}>
+      <div>
         <RegisteredUsersCounter totalUsers={totalUsers} />
-        <div className={classNames.posts}>
-          {items && items.length > 0 ? (
-            items?.map(post => (
-              <div key={post.id}>
-                <Image
-                  alt={'post image'}
-                  height={228}
-                  onClick={() => handlePostClick(post.ownerId, post.id)}
-                  src={post.images[0]?.url}
-                  width={234}
-                />
-                <Link href={`${ROUTES.PROFILE}/${post.ownerId}`}>
-                  <ProfileAvatar avatarOwner={post.avatarOwner} />
-                  <div>{post.userName}</div>
-                </Link>
-                <div>{post.createdAt}</div>
-                <div>{post.description}</div>
-              </div>
-            ))
-          ) : (
-            <div>No posts available</div>
-          )}
-        </div>
+        <PublicPosts items={items} />
       </div>
     </>
   )
