+++ src/shared/ui/index.ts
@@ -6,6 +6,5 @@ export * from './preloader'
 export * from './profileAvatar'
 export * from './profileStat'
 export * from './reCaptcha'
-export * from './registeredUsersCounter'
 export * from './slider'
 export * from './trans'
