--- src/shared/ui/registeredUsersCounter/index.ts
@@ -1,1 +1,0 @@
-export * from './ui'
